Update YA_FSM.cpp
Proposing a change of order of transition checks.

If you for example have two transitions, e.g.

  stateMachine.AddTimedTransition(OPENING, ALARM);
  stateMachine.AddTransition(OPENING, OPENED, [](){return digitalRead(3) == HIGH;});

With the original code, the timed transition will always take precidence and will either go from OPENING -> ALARM or OPENING -> OPENED after x milliseconds depending on the order of definitions of the statemachine transitions. Irrelevent of which one it was meant to go per the timed transition definition.

This happens because the time out state is not tied to the FSM_Transition object, and was instead tied to the State itself, so it was checked and ran on every FSM_Transition first, and will cause a timeout triggered swich to whichever Transition is defined first. 

Ideally timeout could be linked to the FSM_Transition object, but given lack of purpose as far as I can see, for having multiple Timeout transitions on one state (as one would always trigger, the longer one would never) 

I've found this fix to work fine for my application. I don't believe it should cause issues for any other properly implemented systems?